+++ components/createTab/BottomActionsBar.tsx
@@ -8,19 +8,23 @@ import { sliceImage } from "@/components/createTab/utils";
 import { resetImage } from "@/store/slices/imageSlice";
 import store from "@/store";
 
-const ActionsBar = () => {
-  const RunIcon = <Ionicons name="download-outline" size={24} />;
+const BottomActionsBar = () => {
+  const ShareIcon = <Ionicons name="share-outline" size={24} />;
   const ResetIcon = <Ionicons name="refresh-outline" size={24} />;
 
   const handleReset = () => {
     store.dispatch(resetImage());
   };
 
+  const handleShare = () => {
+    sliceImage();
+  };
+
   return (
     <View style={styles.actionsBar} testID="actions-bar">
       <View style={styles.iconsRow}>
         <IconButton icon={ResetIcon} onPress={handleReset} />
-        <IconButton icon={RunIcon} onPress={sliceImage} />
+        <IconButton icon={ShareIcon} onPress={handleShare} />
       </View>
     </View>
   );
@@ -40,4 +44,4 @@ const styles = StyleSheet.create({
   },
 });
 
-export default ActionsBar;
+export default BottomActionsBar;