+++ CHANGELOG.md
@@ -26,6 +26,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Updated project readme
 - Aspect Ratio label becomes a button for cycling through aspect ratio
 - Updated App icon
+- Internally renamed Actions Bar to Bottom Actions Bar
 
 ### Removed